utils modules optimization
