--- javakit/src/main/java/com/dvsnier/utils/AppUtils.java
@@ -1,65 +1,0 @@
-package com.zhy.utils;
-
-import android.content.Context;
-import android.content.pm.PackageInfo;
-import android.content.pm.PackageManager;
-import android.content.pm.PackageManager.NameNotFoundException;
-
-/**
- * 跟App相关的辅助类
- * 
- * @author zhy
- * 
- */
-public class AppUtils
-{
-
-	private AppUtils()
-	{
-		/* cannot be instantiated */
-		throw new UnsupportedOperationException("cannot be instantiated");
-
-	}
-
-	/**
-	 * 获取应用程序名称
-	 */
-	public static String getAppName(Context context)
-	{
-		try
-		{
-			PackageManager packageManager = context.getPackageManager();
-			PackageInfo packageInfo = packageManager.getPackageInfo(
-					context.getPackageName(), 0);
-			int labelRes = packageInfo.applicationInfo.labelRes;
-			return context.getResources().getString(labelRes);
-		} catch (NameNotFoundException e)
-		{
-			e.printStackTrace();
-		}
-		return null;
-	}
-
-	/**
-	 * [获取应用程序版本名称信息]
-	 * 
-	 * @param context
-	 * @return 当前应用的版本名称
-	 */
-	public static String getVersionName(Context context)
-	{
-		try
-		{
-			PackageManager packageManager = context.getPackageManager();
-			PackageInfo packageInfo = packageManager.getPackageInfo(
-					context.getPackageName(), 0);
-			return packageInfo.versionName;
-
-		} catch (NameNotFoundException e)
-		{
-			e.printStackTrace();
-		}
-		return null;
-	}
-
-}
